pyproject.toml
[project]
name = "pdf-book-configure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.20.0",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.6",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
